requirements.txt
alembic==1.7.7
apturl==0.5.2
asgiref==3.5.2
attrs==21.4.0
Babel==2.10.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
black==22.3.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==4.0.6
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.7.0
flake8==4.0.1
flake8-polyfill==1.0.2
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
huepy==1.2.1
idna==2.8
imageio==2.16.2
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
instabot==0.117.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
kazam==1.4.5
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mock==4.0.3
monotonic==1.5
moviepy==1.0.3
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pathspec==0.9.0
pep8==1.7.1
pep8-naming==0.12.1
pexpect==4.6.0
Pillow==9.1.0
platformdirs==2.5.2
pluggy==1.0.0
postgres==4.0
proglog==0.1.9
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
py==1.11.0
pycairo==1.16.2
pycodestyle==2.8.0
pycups==1.9.73
pyflakes==2.4.0
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.8
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.9
pytest==7.1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.20.0
schedule==1.1.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
systemd-python==234
texttable==1.6.4
tomli==2.0.1
tqdm==4.64.0
tweepy==4.10.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.14.1
wadllib==1.3.3
websocket-client==0.59.0
Werkzeug==2.1.2
whitenoise==6.2.0
xkit==0.0.0
zipp==3.8.0
